requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Pillow==10.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
img2pdf==0.5.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.5
psutil==5.9.8
